+++ changelog.md
@@ -5,6 +5,8 @@
 > [!IMPORTANT]
 > This version is not released yet and is under active development.
 
+- [mpm] Remap pURL types to managers. Closes {issue}`1460`.
+
 ## [5.21.0 (2025-05-28)](https://github.com/kdeldycke/meta-package-manager/compare/v5.20.0...v5.21.0)
 
 - [mpm] Drop support for Python 3.10.